chore: Add release workflow and update dependencies
This commit adds a new release workflow in the `.github/workflows/release.yml` file. The workflow is triggered on push events with tags matching the pattern 'v*'. It creates a release, builds the project, zips the release, and uploads it as a release asset. Additionally, the major version tag is updated.

The `package-lock.json` and `package.json` files were also modified to update dependencies. The `@types/rollup` dependency was removed, and the `picomatch` dependency was added.